Fixed errors.
Initialization of sentinel nodes (A.K.A. head and tail) moved from NULL to new Node<T>.

Deleted declarations of unused methods.

Insertion operator overloaded function ( operator<<() ) moved from protected to public field.

Additive assignment operator ( a += b ) overloaded so that the second collection is concatenated to the first, appending the head node of 'b' to the tail node of 'a' and so forth. This operation replaces the 'append' method.

Created 'linkSentinels' method, intended for the cases when the collection is empty, this means the 'head' node will point to the 'tail' node both forward and backward.

Check for collection size is made now by looking at the length of it instead of checking nullity over sentinel nodes. This due to sentinel nodes never being null.